Change code places
